backend/requirements.txt
Django==3.2
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-extra-fields==3.7.0
djoser==2.1.0
PyJWT==2.1.0
psycopg2-binary==2.9.3
Pillow==9.3.0
python-dotenv==1.0.1